Copy utility functions to package

This function copies handy utility functions to a new script in a specified location. Currently, the script contains the following utility functions:

.assertScalar() - convenience function to check the validity of scalar variables.

.assertVector() - convenience function to check the validity of vector variables.

addUtilsFunctions(outFile = "R/utils.R", copyTests = TRUE)

Arguments

outFile

Character scalar, giving the path to which the script should be copied. The path is relative to the root of the active project. If a file with this name already exists, the function will ask for confirmation before overwriting it.

copyTests

Logical scalar, defining whether to copy unit tests for the utility functions to tests/testthat/test-<name>.R, where <name> is the base name of outFile. If the target package is not yet set up to use testthat, the function will also run usethis::use_testthat() to generate the required folder structure and add testthat to the list of suggested package in the DESCRIPTION file.

Author

Charlotte Soneson