Plot annotated phasogram.

Plot phasogram and annotate it with estimated nucleosome repeat length (NRL).

plotPhasogram(x, hide = TRUE, xlim = NULL, verbosePlot = FALSE, ...)

Arguments

x
numeric vector giving the counts of alignment distances (typically the output of calcPhasogram.

hide
If TRUE (the default), hide phasogram counts not used in the NRL estimate (mind parameter from estimateNRL).

xlim
numeric(2) with the x-axis (phase) limits in the first two plots (see Details). if NULL (the default), the full range defined by x and hide will be used.

verbosePlot
If TRUE, create three plots instead of just a single plot (see Details).

...
Additional arguments passed to estimateNRL to control NRL estimation.

Value

The return value from the call to estimateNRL (invisibly).

Details

The function will visualize an annotated phasogram. For verbosePlot=FALSE (the default), it will create a single annotated plot. For verbosePlot=TRUE, it will create three plots (using par(mfrow=c(1,3))):

raw phase counts with de-trending and de-noising loess fits

residual phase counts with de-noising loess fit and detected peaks

linear fit to peaks and NRL estimation

See also

Examples

  # see the help for calcPhasogram() for a full example